Move the histogramming code to parse directory
