+++ link-grammar/api.c
@@ -23,8 +23,8 @@
 #include "disjunct-utils.h"  // for free_disjuncts
 #include "error.h"
 #include "externs.h"
-#include "histogram.h"
 #include "linkage.h"
+#include "parse/histogram.h"  // for PARSE_NUM_OVERFLOW
 #include "parse/parse.h"
 #include "post-process/post-process.h" // for linkage_free_pp_info
 #include "print/print.h"
