+++ link-grammar/parse/prune.c
@@ -13,8 +13,6 @@
 
 #include "api-structures.h"
 #include "connectors.h"
-// #include "dict-common/dict-api.h"   /* for print_expression when debugging */
-// #include "dict-common/dict-utils.h" /* for free_E_list */
 #include "disjunct-utils.h"
 #include "post-process/post-process.h"
 #include "post-process/pp-structures.h"
